fix(rust): rename the kid_uri to just an id_uri to better reflect its type
