+++ rust/catalyst-types/src/id_uri/errors.rs
