+++ rust/catalyst-types/src/id_uri/key_rotation.rs
