+++ rust/catalyst-types/src/id_uri/role_index.rs
