+++ rust/catalyst-types/src/lib.rs
@@ -2,6 +2,6 @@
 
 pub mod conversion;
 pub mod hashes;
-pub mod kid_uri;
+pub mod id_uri;
 pub mod problem_report;
 pub mod uuid;
