+++ rust/catalyst-types/tests/mod.rs
@@ -12,7 +12,7 @@ fn test_type_usage() {
     type D = catalyst_types::uuid::V4;
     type E = catalyst_types::uuid::V7;
 
-    type F = catalyst_types::kid_uri::KidUri;
+    type F = catalyst_types::id_uri::KidUri;
 
     let bytes: [u8; 32] = [0; 32];
     let _ = catalyst_types::hashes::Blake2bHash::from(bytes);
